refactor(worker): extract the API HTTP client into api_client.rs (sc-1636)
Move the ApiClient struct/impl and the decode_api_response helper out of
lib.rs into api_client.rs. Settings/WorkerError/WorkerResult stay at the
root as shared core types (resolved in the module via `use super::*`).

pub(crate) on ApiClient + new/get_json/post_json (used by the run loop
and job handlers); url/with_auth/decode_api_response stay private. Worker
tests.rs ApiClient import repointed to super::api_client.
Behavior-preserving. lib.rs 4,415 -> 4,350 lines.
