+++ crates/sceneworks-worker/src/lib.rs
@@ -31,6 +31,8 @@ use tokio::process::{Child, Command};
 use tokio::time::MissedTickBehavior;
 use uuid::Uuid;
 
+mod api_client;
+use api_client::*;
 mod gpu;
 use gpu::*;
 mod supervisor;
@@ -176,72 +178,6 @@ impl From<ProjectStoreError> for WorkerError {
 }
 
 pub type WorkerResult<T> = Result<T, WorkerError>;
-
-#[derive(Clone)]
-struct ApiClient {
-    client: reqwest::Client,
-    api_url: String,
-    access_token: Option<String>,
-}
-
-impl ApiClient {
-    fn new(settings: &Settings) -> Self {
-        Self {
-            client: reqwest::Client::new(),
-            api_url: settings.api_url.trim_end_matches('/').to_owned(),
-            access_token: settings.access_token.clone(),
-        }
-    }
-
-    async fn get_json<T>(&self, path: &str) -> WorkerResult<T>
-    where
-        T: for<'de> Deserialize<'de>,
-    {
-        let response = self
-            .with_auth(self.client.get(self.url(path)))
-            .send()
-            .await?;
-        decode_api_response(response).await
-    }
-
-    async fn post_json<T, U>(&self, path: &str, payload: &T) -> WorkerResult<U>
-    where
-        T: serde::Serialize + ?Sized,
-        U: for<'de> Deserialize<'de>,
-    {
-        let response = self
-            .with_auth(self.client.post(self.url(path)).json(payload))
-            .send()
-            .await?;
-        decode_api_response(response).await
-    }
-
-    fn url(&self, path: &str) -> String {
-        format!("{}{}", self.api_url, path)
-    }
-
-    fn with_auth(&self, request: reqwest::RequestBuilder) -> reqwest::RequestBuilder {
-        match &self.access_token {
-            Some(token) => request.header("X-SceneWorks-Token", token),
-            None => request,
-        }
-    }
-}
-
-async fn decode_api_response<T>(response: reqwest::Response) -> WorkerResult<T>
-where
-    T: for<'de> Deserialize<'de>,
-{
-    let status = response.status();
-    if !status.is_success() {
-        let detail = response
-            .text()
-            .await
-            .unwrap_or_else(|_| "request failed".to_owned());
-        return Err(WorkerError::Api { status, detail });
-    }
-    Ok(response.json::<T>().await?)
-}
 
 #[derive(Debug, Clone, PartialEq)]
 struct DiscoveredGpu {
