+++ crates/sceneworks-worker/src/tests.rs
@@ -13,6 +13,7 @@ use sceneworks_core::contracts::WorkerUtilizationSnapshot;
 use serde_json::{json, Value};
 use tempfile::tempdir;
 
+use super::api_client::ApiClient;
 use super::gpu::{
     cpu_gpu, cpu_worker_id, fallback_gpu, gpu_worker_id, parse_nvidia_smi_gpus, visible_gpu_ids,
     worker_capabilities_with_utility,
@@ -27,10 +28,10 @@ use super::{
     download_lora_source_url, download_progress_payload, download_snapshot, finalize_converted_dir,
     fresh_asset_id, import_lora_source_path, now_rfc3339, output_dimensions,
     resolve_model_convert_output, resolve_model_import_target, run_ffmpeg, safe_download_dir,
-    safe_project_path, value_f64, write_model_install_marker, ApiClient, DownloadContext,
-    DownloadFamilyCheck, DownloadProgress, HuggingFaceSnapshot, JsonObject, Settings, SnapshotFile,
-    WorkerError, DEFAULT_MAX_LORA_URL_BYTES, DEFAULT_MAX_MODEL_URL_BYTES,
-    DEFAULT_TRANSITION_DURATION_SECONDS, INSTALL_MARKER,
+    safe_project_path, value_f64, write_model_install_marker, DownloadContext, DownloadFamilyCheck,
+    DownloadProgress, HuggingFaceSnapshot, JsonObject, Settings, SnapshotFile, WorkerError,
+    DEFAULT_MAX_LORA_URL_BYTES, DEFAULT_MAX_MODEL_URL_BYTES, DEFAULT_TRANSITION_DURATION_SECONDS,
+    INSTALL_MARKER,
 };
 
 fn write_safetensors_with_keys(path: &std::path::Path, keys: &[String]) {
